Funcionamiento modulo registro de Pacas
Se cambio el nombre del archivo para el registro de Pacas y se implemento el funcionamiento para agregar al Cart las pacas
