+++ app_inventory/lib/menu.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'add.dart';
+import 'paca_register.dart';
 import 'home.dart';
 
 class Menu extends StatefulWidget {
@@ -13,7 +13,7 @@ class Menu extends StatefulWidget {
 class _Menu extends State<Menu> {
   int indexTap = 1;
   var titleBar = "Inventario";
-  final List<Widget> widgetsChildren = [Add(), Home(), Add()];
+  final List<Widget> widgetsChildren = [PacaRegister(), Home(), PacaRegister()];
   void onTapTapped(int index) {
     setState(() {
       indexTap = index;
@@ -43,8 +43,7 @@ class _Menu extends State<Menu> {
       body: widgetsChildren[indexTap],
       backgroundColor: Color(0xffefefef),
       bottomNavigationBar: Theme(
-        data: Theme.of(context)
-            .copyWith(canvasColor: Colors.white),
+        data: Theme.of(context).copyWith(canvasColor: Colors.white),
         child: BottomNavigationBar(
           onTap: onTapTapped,
           currentIndex: indexTap,
